+++ imports/charts/client/index.js
@@ -1,10 +1,2 @@
+import '../charts';
 import './message-day-chart';
-
-ChartsController = ApplicationController.extend({
-  template: 'MessageDayChart'
-});
-
-Router.route('/charts', {
-  name: 'charts',
-  controller: 'ChartsController'
-});