+++ imports/charts/client/message-day-chart.js
@@ -3,7 +3,7 @@ import './message-day-chart.html';
 import Chart from 'chart.js';
 import moment from 'moment';
 
-Template.MessageDayChart.onRendered(function() {
+Template.MessagesDailyChart.onRendered(function() {
   let days = moment().endOf("month").date();
   Meteor.call("getChartDay", (err, res) => {
     if (err) {
@@ -15,17 +15,17 @@ Template.MessageDayChart.onRendered(function() {
         data: {
           labels: _.range(1, days),
           datasets: [{
-            label: '内部',
+            label: I18n.t('Internal'),
             data: res["internal"],
             backgroundColor: 'rgba(200, 200, 200, 0.8)',
             borderWidth: 1
           }, {
-            label: '发出',
+            label: I18n.t('Outgoing'),
             data: res["outgoing"],
             backgroundColor: 'rgba(0, 0, 255, 0.8)',
             borderWidth: 1
           }, {
-            label: '接收',
+            label: I18n.t('Incoming'),
             data: res["incoming"],
             backgroundColor: 'rgba(0, 128, 0, 0.8)',
             borderWidth: 1
@@ -48,7 +48,10 @@ Template.MessageDayChart.onRendered(function() {
       });
     }
   });
+});
 
+Template.MessagesHourlyChart.onRendered(function() {
+  let days = moment().endOf("month").date();
   Meteor.call("getCharHour", (err, res) => {
     if (err) {
       console.log(err);
@@ -62,17 +65,17 @@ Template.MessageDayChart.onRendered(function() {
         data: {
           labels: _.range(1, 24),
           datasets: [{
-            label: '内部',
+            label: I18n.t('Internal'),
             data: res["internal"],
             backgroundColor: 'rgba(200, 200, 200, 0.8)',
             borderWidth: 1
           }, {
-            label: '发出',
+            label: I18n.t('Outgoing'),
             data: res["outgoing"],
             backgroundColor: 'rgba(0, 0, 255, 0.8)',
             borderWidth: 1
           }, {
-            label: '接收',
+            label: I18n.t('Incoming'),
             data: res["incoming"],
             backgroundColor: 'rgba(0, 128, 0, 0.8)',
             borderWidth: 1