+++ imports/i18n/locales/en-US.js
@@ -11,5 +11,9 @@ export const locale_en_us = {
   "Created by and when": "Created by{{name}} at {{time}}",
   "mail to": "to",
   "mail cc": "cc",
-  "confirm remove member": "Are you sure to remove {{name}}?"
+  "confirm remove member": "Are you sure to remove {{name}}?",
+  // reports
+  "System Reports": "Reports",
+  "Daily Messages Counts": "Daily",
+  "Hourly Messages Counts": "Hourly"
 };